rename files in variational
