+++ netket/variational/mc_mixed_state.py
